refactor: move `send_email` back to services
